rename account profile to reset name
